backend/requirements.txt
flask
flask-cors
python-dotenv
bcrypt
google-genai>=1.0.0
psutil>=5.9.0